# internal/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nops API",
    "version": "1",
    "license": { "name": "Apache 2.0", "identifier": "Apache-2.0" },
    "description": "Read jobs and deployments, act on them, and manage ACL policies, binding rules and tokens. A token does what its ACL policies allow, with the dashboard's rules."
  },
  "servers": [{ "url": ".." }],
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/api/acl/policies/{name}": {
      "get": {
        "summary": "One ACL policy with its rules",
        "operationId": "getACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "responses": {
          "200": {
            "description": "The ACL policy.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicy" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "summary": "Create an ACL policy or replace it",
        "operationId": "putACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicyBody" } } }
        },
        "responses": {
          "200": {
            "description": "The ACL policy as saved.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicy" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Delete an ACL policy",
        "operationId": "deleteACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/acl/binding-rules/{id}": {
      "get": {
        "summary": "One binding rule",
        "operationId": "getBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "responses": {
          "200": {
            "description": "The binding rule.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRule" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "summary": "Replace a binding rule",
        "operationId": "putBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRuleBody" } } }
        },
        "responses": {
          "200": {
            "description": "The binding rule as saved.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRule" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Delete a binding rule",
        "operationId": "deleteBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/acl/tokens/{accessor_id}": {
      "get": {
        "summary": "One token",
        "operationId": "getACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "200": {
            "description": "The token.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLToken" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Revoke a token",
        "operationId": "revokeACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "204": { "description": "Revoked." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "A token from the dashboard's Administration page." }
    },
    "parameters": {
...
      "PolicyName": { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
      "AccessorID": { "name": "accessor_id", "in": "path", "required": true, "schema": { "type": "string" } },
      "BindingRuleID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is not what the endpoint takes.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "No valid token: it is missing, unknown, expired, or revoked.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The token is valid, and its ACL policies do not allow this.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The job, deployment, ACL policy or token does not exist.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ServerError": {
        "description": "Nops hit an error and logged it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "ACLPolicy": {
        "type": "object",
        "required": ["name", "description", "rules", "created_at", "modified_at"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "rules": { "type": "string", "description": "The rules, in HCL." },
          "created_at": { "type": "string", "format": "date-time" },
          "modified_at": { "type": "string", "format": "date-time" }
        }
      },
      "ACLPolicyBody": {
        "type": "object",
        "required": ["rules"],
        "properties": {
          "description": { "type": "string" },
          "rules": { "type": "string", "description": "The rules, in HCL. Rules that do not parse are refused and nothing is saved." }
        }
      },
      "BindingRule": {
        "type": "object",
        "required": ["id", "description", "auth_method", "selector", "bind_type", "created_at", "modified_at"],
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string" },
          "auth_method": { "type": "string", "enum": ["oidc", "basic"] },
          "selector": { "type": "string", "description": "A go-bexpr expression on the claims of a login: value.username, value.sub and list.groups. Empty matches every login of the auth method." },
          "bind_type": { "type": "string", "enum": ["policy", "management"] },
          "bind_name": { "type": "string", "description": "The ACL policy a policy rule binds. Absent for management." },
          "created_at": { "type": "string", "format": "date-time" },
          "modified_at": { "type": "string", "format": "date-time" }
        }
      },
      "BindingRuleBody": {
        "type": "object",
        "required": ["auth_method", "bind_type"],
        "properties": {
          "description": { "type": "string", "maxLength": 200 },
          "auth_method": { "type": "string", "enum": ["oidc", "basic"] },
          "selector": { "type": "string", "maxLength": 1024, "description": "Refused if it does not parse or reads a claim that does not exist." },
          "bind_type": { "type": "string", "enum": ["policy", "management"] },
          "bind_name": { "type": "string", "description": "The ACL policy to bind, which must exist. Empty for management." }
        }
      },
      "ACLToken": {
        "type": "object",
        "required": ["accessor_id", "name", "type", "origin", "policies", "created_at"],
        "properties": {
          "accessor_id": { "type": "string", "description": "Public: it names the token in logs and the audit, and grants nothing." },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["management", "client"] },
          "origin": { "type": "string", "enum": ["login", "created"], "description": "A login made a session token; a person or a token created the others." },
          "identity": { "type": "string", "description": "The person a session token stands for. Absent for a created token." },
          "policies": { "type": "array", "items": { "type": "string" }, "description": "The ACL policies a client token carries, by name." },
          "created_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time", "description": "Absent for a token that never expires." },
          "creator_accessor_id": { "type": "string", "description": "The token that made this one. It stays after that token is gone." },
          "creator_name": { "type": "string" },
          "creator_identity": { "type": "string", "description": "The person who created the token, when it was one." }
        }
      },
...
    }
  }
}